+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Vue 2.5+ is now the minimum required version
 * **Removed** prompt dialog ``inputPlaceholder``, ``inputName`` and ``inputMaxlength`` props
+* Add ``disabled`` prop to tabItem
 
 ## 0.5.4